Don't use homebrew golangci-lint
This sounds mad, but sometimes, when using a version of golangci-lint
that was compiled with a different version of the go compiler to what
I'm using on a project, it balloons in memory usage to >100GB of virtual
memory on certain projects. I have no idea why this fixes it, but
ensuring I'm using a go-installed, locally-compiled, farm-to-table
version of golangci-lint solves the issue. Well, that and/or Docker.